cleanup(core): extract watcher event routing into a reusable helper
handleWorkspaceChanges in server.ts mixed daemon-process bookkeeping
(inactivity timer, error tracking) with the actual event routing
(stat-filter, categorize, schedule recompute). The routing portion
is the part that matters for tests; the bookkeeping is daemon-only
state.

Pulls routeWorkspaceChanges into file-watching/ so the daemon and
the integration spec both use the same routing path. The daemon
keeps its bookkeeping wrapper; the test uses a no-op callback that
forwards events straight to the helper.

No behavioural change.
